refactor: remove duplicate toast hook
