+++ app/adminAZ/certificates/edit/[id]/page.tsx
@@ -30,7 +30,7 @@ import {
   PopoverTrigger,
 } from "@/components/ui/popover";
 import { Textarea } from "@/components/ui/textarea";
-import { useToast } from "@/components/ui/use-toast";
+import { useToast } from "@/hooks/use-toast";
 import { cn } from "@/lib/utils";
 import { FadeIn } from "@/components/animations/fade-in";
 import { updateService, getServiceById } from "@/lib/api-services";
